feat:初步完成容器性能查询接口

+++ src/main/java/org/example/performance/pojo/po/ContainerMetrics.java
@@ -34,7 +34,7 @@ public class ContainerMetrics implements Serializable {
     /**
      * CPU使用率
      */
-    private Double cpuRate;
+    private BigDecimal cpuRate;
 
     /**
      * 重启时间
